+++ app/Http/Livewire/Layanan/LayanansTable.php
@@ -3,12 +3,15 @@
 namespace App\Http\Livewire\Layanan;
 
 use App\Models\Layanans;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Livewire\WithPagination;
 
 class LayanansTable extends Component
 {
+    
    use WithPagination;
+   use AuthorizesRequests;
     protected $paginationTheme = 'bootstrap';
 
     public $layanan_id;
@@ -19,6 +22,7 @@ class LayanansTable extends Component
 
     public function render()
     {
+        $this->authorize('adminview', Layanans::class);
         return view('livewire.layanan.layanans-table',[
             'layanans' => Layanans::orderBy('id', 'desc')->paginate(5)
         ]);
@@ -26,12 +30,14 @@ class LayanansTable extends Component
 
     public function layananEdit($layanan)
     {
+        $this->authorize('adminview', Layanans::class);
         $this->layanan_id = $layanan['id'];
         $this->nama_layanan = $layanan['nama_layanan'];
     }
 
     public function layananUpdate()
     {
+        $this->authorize('adminview', Layanans::class);
         $this->validate([
             'nama_layanan' => 'required',
         ]);
@@ -48,6 +54,7 @@ class LayanansTable extends Component
 
     public function layananDelete($id)
     {
+        $this->authorize('adminview', Layanans::class);
         $this->layanan_id = $id;
         
         $layanan = Layanans::find($id);
@@ -56,6 +63,7 @@ class LayanansTable extends Component
 
     public function layananDestroyer()
     {
+        $this->authorize('adminview', Layanans::class);
         Layanans::find($this->layanan_id)->delete();
     }
 }
